Add a stdlib module
This provides a place to define standard library functions. Currently,
it provides the type environment and overloading assumptions for
several primitive numerical functions, which can be used to test type
inference in more complex programs.

One of the tests fails because instantiation of quantified type
variables does not respect the overloading assumption.