Use the standard library types in the repl/run functions
This clears the type substitution after every type checking, but
requires us to be careful that we don't overwrite the fresh type
variables associated with the standard library functions. We avoid
this by using fixed type variables that do not coincide with the fresh
variables in the standard library type environment.