Create a parent pointer tree for the type environments
Since type environments are immutable, it suffices to use an Rc
pointer to refer to the parent environment. There are a few patterns,
namely when we destroy the local type environments after checking
the bodies of lambda terms and in the REPL when we insert newly
defined procedures into the type environment that we need to be a
little creative in extracting the environment from Rc pointers. We
should only ever have one reference to the particular environment in
these moments, so we either `try_unwrap` or `get_mut` the Rc and panic
if it fails.