Add floating point literal overloading
This uses an approach similar to Haskell where floating point literals
are converted into calls to the function fromFloating, which is
overloaded with implementations for the two floating point number
types (fromFloatingf32 and fromFloatingf64).

Haskell uses a fromRational conversion from an arbitrary precision
rational number, but that requires slightly more sophisticated parsing
of floating point literals into rational numbers.

We can also use the same approach with fromInteger, but that change is
slightly more intrusive, and it is put off for later.

I am not entirely sure that the literal overloading is resolved
correctly in either the compiler or the interpreter and more tests are
necessary.