Add some functions to force floating point types
The overloading resolution test failed because it failed to specialize
the function for a specific floating point, which then got eliminated
in the resolution pass. With some additional functions, we can force
floating point literals into certain types. This is hacky, but it is
necessary to avoid too much ambiguity.

In practice, the data that we operate on will be fixed to a particular
floating point type, and the overloading resolution pass should coerce
all the literals to the correct types.