Stub out a let-desugaring pass
This will convert the program into a single expression after type
checking, so overloading resolution, alpha conversion, A-normalization
and so on operate on a single expression. We modify Program to ensure
that it consists of a set of definitions followed by an
expression. Compilation currently fails because let desugaring is not
completely implemented.